src/requirements.txt
Django==4.1.1
djangorestframework==3.12.4
Pillow==8.3.1
pytz==2023.2
djangorestframework-simplejwt==5.2.2
drf-yasg==1.20.0
tzdata==2023.3
typing-extensions==4.7.1

